+++ front/lib/core/utils/app_pages.dart
@@ -1,5 +1,5 @@
 // import 'package:get/get_navigation/src/routes/get_route.dart';
-// import 'package:idekiller/presentation/home_page.dart';
+// import 'package:idekiller/presentation/home_screen.dart';
 // import 'package:idekiller/screens/auth/auth_screen.dart';
 // import 'package:idekiller/screens/compiler.dart';
 // import 'package:idekiller/screens/user_profile.dart';
