+++ front/lib/core/utils/compiler_classes/loading_animation.dart
