+++ front/lib/core/utils/initial_codes.dart
@@ -1,0 +1,16 @@
+Map<String, String> initialCodes = {
+  "C++": "#include <iostream>\n"
+      "\n"
+      "using namespace std;\n"
+      "int main(void) {\n"
+      "  cout << \"Hello world\" <<endl;\n"
+      "}\n",
+  "Java":
+      "public class Main {\n  public static void main (String[] args) {\n    System.out.println(\"Hello, World\");\n  }\n}",
+  "C": "#include <stdio.h>\n"
+      "\n"
+      "int main(void) {\n"
+      "  printf(\"Hello world\");\n"
+      "}\n",
+  "Python": "print(\"Hello, World!\")"
+};
