+++ front/lib/features/compiler/presentation/bloc/code_bloc_event.dart
@@ -1,0 +1,9 @@
+sealed class CodeBlocEvent {}
+
+class CodeBlocEventFetch extends CodeBlocEvent {}
+class CodeBlocEventLanguageChangeJava extends CodeBlocEvent {}
+class CodeBlocEventLanguageChangePython extends CodeBlocEvent {}
+class CodeBlocEventLanguageChangeCpp extends CodeBlocEvent {}
+class CodeBlocEventLanguageChangeC extends CodeBlocEvent {}
+
+
