--- front/lib/utils/global_values.dart
@@ -1,10 +1,0 @@
-import 'package:idekiller/controllers/model.dart';
-
-class GlobalValues {
-  static String language = 'Java';
-  static double font = 14;
-  static String code = "public class Main {\n  public static void main (String[] args) {\n    System.out.println(\"Hello, World\");\n  }\n}";
-  static User user = User(profilePic: '', name: 'Masss', number: '123', students: [], teachers: [], id: 0, about: 'qwe rty');
-  static bool isAuthorized = false;
-  static String uuid = "";
-}
