+++ front/pubspec.yaml
@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.17.0 <4.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -26,6 +26,7 @@ dependencies:
   url_launcher: ^6.1.10
   uuid: ^4.5.0
   go_router: ^14.2.7
+  flutter_bloc: ^8.1.6
 
 dev_dependencies:
   flutter_test:
